Move tests to test_capi.test_codecs.
